ED13846 fixed handler and scripts initialization
